selftests/mm: static process_madvise() wrapper for guard-pages
The recently introduced guard-pages mm selftest uses the process_madvise()
syscall, a wrapper for which was added to glibc v2.36.  For those of us
stuck with older distributions this causes a compile error when compiling
the mm selftests.  For example Ubuntu 22.04 uses glibc 2.35, which does
not have the wrapper.

To workaround the issue, let's introduce our own static process_madvise()
wrapper that uses glibc's syscall() helper.

While we are at it, add the guard-page test suite to run_vmtests.sh so
that it can be automatically run by CI systems.

Link: https://lkml.kernel.org/r/20250107142937.1870478-1-ryan.roberts@arm.com
Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>
Reviewed-by: Lorenzo Stoakes <lorenzo.stoakes@oracle.com>
Cc: Jann Horn <jannh@google.com>
Cc: Liam R. Howlett <Liam.Howlett@Oracle.com>
Cc: Shuah Khan <shuah@kernel.org>
Cc: Vlastimil Babka <vbabka@suse.cz>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>