mm/damon/core: add damos_filter->pass field
DAMOS filters work as only exclusive (block) filters.  This makes it easy
to be confused, and restrictive at combining multiple filters for various
types of memory.

Add a field anmed 'pass' to damos_filter.  The field will be used to
indicate whether the filter should work for inclusion or not.

Following two commits will make the core and operations set layer, which
handles the damos_filter object, respect the field, respectively.

Link: https://lkml.kernel.org/r/20250107201739.79484-3-sj@kernel.org
Signed-off-by: SeongJae Park <sj@kernel.org>
Cc: Jonathan Corbet <corbet@lwn.net>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>