mm/shmem: fix invalid PTR_ERR(NULL) call in shmem_xattr_handler_set()
In shmem_xattr_handler_set() if simple_xattr_set() succeeds and the
pointer returned is not an error pointer, then old_xattr will be set to
NULL in the body of the following if statement:

    if (!IS_ERR(old_xattr))

Later on shmem_xattr_handler_set() calls:

    return PTR_ERR(old_xattr);

The PTR_ERR macro is used to extract an error code from an error pointer
and NULL is not an error pointer, PTR_ERR(NULL) simply results in 0.

To improve correctness and readability, refactor the error handling to
have an explicit default return value of 0 (success) in "ret".  If
simple_xattr_set() returns an error pointer, store its error code in
"ret".

Link: https://lkml.kernel.org/r/20250128235408.11229-1-qasdev00@gmail.com
Signed-off-by: Qasim Ijaz <qasdev00@gmail.com>
Cc: Hugh Dickens <hughd@google.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>